fix: Исправить код-стайл в InMemoryHistoryManager, Реворк класса Node, Добавить класс DoubleLinkedList
